WLT-297: renamed to crypto and xcrypto
