+++ platformpolicy/crypto/internal/wrapper_test.go
@@ -17,14 +17,14 @@
 package internal
 
 import (
-	"github.com/insolar/insolar/platformpolicy/commoncrypto"
+	"github.com/insolar/insolar/platformpolicy/crypto"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 
 	"github.com/insolar/insolar/component"
-	"github.com/insolar/insolar/platformpolicy/commoncrypto/internal/hash"
-	"github.com/insolar/insolar/platformpolicy/commoncrypto/internal/sign"
+	"github.com/insolar/insolar/platformpolicy/crypto/internal/hash"
+	"github.com/insolar/insolar/platformpolicy/crypto/internal/sign"
 	"github.com/insolar/insolar/testutils"
 )
 
@@ -32,7 +32,7 @@ func TestEcdsaMarshalUnmarshal(t *testing.T) {
 	count := 10000
 	data := testutils.RandomRef()
 
-	kp := commoncrypto.NewKeyProcessor()
+	kp := crypto.NewKeyProcessor()
 	provider := sign.NewECDSAProvider()
 
 	cm := component.Manager{}
