+++ platformpolicy/xcrypto/cryptographyscheme.go
@@ -14,15 +14,15 @@
 // limitations under the License.
 //
 
-package customcrypto
+package xcrypto
 
 import (
 	"github.com/insolar/insolar/component"
 	"github.com/insolar/insolar/insolar"
 	"github.com/insolar/insolar/platformpolicy/algorithmprovider"
-	"github.com/insolar/insolar/platformpolicy/customcrypto/internal/hash"
-	"github.com/insolar/insolar/platformpolicy/customcrypto/internal/sign"
 	"github.com/insolar/insolar/platformpolicy/keys"
+	"github.com/insolar/insolar/platformpolicy/xcrypto/internal/hash"
+	"github.com/insolar/insolar/platformpolicy/xcrypto/internal/sign"
 )
 
 type platformCryptographyScheme struct {
