+++ platformpolicy/xcrypto/cryptographyscheme_test.go
@@ -14,7 +14,7 @@
 // limitations under the License.
 //
 
-package customcrypto
+package xcrypto
 
 import (
 	"testing"
